rename self_build to build
